services/auth0/contracts/pyproject.toml
[project]
name = "auth0-contracts"
version = "1.0.0"
description = "Contract tests for Auth0 fake"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0",
    "httpx>=0.27",
    "auth0-python>=4.7",
    "pyjwt[crypto]>=2.8.0",
]
